Edit Dashboard.jsx to import the new SidebarLayout instead of having it's own sidebar
